+++ src/transport/Car.java
@@ -1,12 +1,9 @@
 package transport;
 
-public class Car {
-    private final String brand;
-    private final String model;
+import static transport.Validate.*;
+
+public class Car extends Transport {
     private double engineVolume; //объем двигателя в литрах
-    private String color;
-    private final int productionYear;
-    private final String productionCountry;
     private String transmission;
     private final String bodyType;
     private String registrationNumber;
@@ -26,12 +23,8 @@ public class Car {
                int numberOfSeats,
                boolean isSummerTires,
                Key key) {
-        this.brand = validateCarParameters(brand);
-        this.model = validateCarParameters(model);
+        super(brand,model,productionYear,productionCountry,color,0);
         this.engineVolume = validateCarEngineVolume(engineVolume);
-        this.color = validateCarColor(color);
-        this.productionYear = validateCarProductionYear(productionYear);
-        this.productionCountry = validateCarParameters(productionCountry);
         this.transmission = validateCarTransmission(transmission);
         this.bodyType = validateCarBodyType(bodyType);
         this.registrationNumber = validateCarRegistrationNumber(registrationNumber);
@@ -39,29 +32,8 @@ public class Car {
         this.isSummerTires = isSummerTires;
         this.key = key;
     }
-
-    public String getBrand() {
-        return brand;
-    }
-
-    public String getModel() {
-        return model;
-    }
-
     public double getEngineVolume() {
         return engineVolume;
-    }
-
-    public String getColor() {
-        return color;
-    }
-
-    public int getProductionYear() {
-        return productionYear;
-    }
-
-    public String getProductionCountry() {
-        return productionCountry;
     }
 
     public String getTransmission() {
@@ -91,10 +63,6 @@ public class Car {
         this.engineVolume = validateCarEngineVolume(engineVolume);
     }
 
-    public void setColor(String color) {
-        this.color = validateCarColor(color);
-    }
-
     public void setTransmission(String transmission) {
         this.transmission = validateCarTransmission(transmission);
     }
@@ -109,9 +77,9 @@ public class Car {
 
     @Override
     public String toString() {
-        return brand + ' ' + model + ", страна сборки - " + productionCountry +
-                ", год выпуска - " + productionYear +
-                ",\n\t цвет кузова - " + color +
+        return getBrand() + ' ' + getModel() + ", страна сборки - " + getProductionCountry() +
+                ", год выпуска - " + getProductionYear() +
+                ",\n\t цвет кузова - " + getColor() +
                 ",\n\t объем двигателя - " + engineVolume + " л." +
                 ",\n\t коробка передач - " + transmission +
                 ",\n\t тип кузова - " + bodyType +
@@ -121,44 +89,7 @@ public class Car {
                 ",\n\t " + key;
     }
 
-    public static String validateCarParameters(String value){
-        if (value == null || value.isBlank()) {return "default";}
-        return value;
-    }
-    public static String validateCarColor(String value) {
-        if (value == null || value.isBlank()) {return "белый";}
-        return value;
-    }
-    public static String validateCarTransmission(String value){
-        if (value == null || value.isBlank()) {return "автомат";}
-        return value;
-    }
-    public static String validateCarBodyType(String value){
-        if (value == null || value.isBlank()) {return "седан";}
-        return value;
-    }
-    public static double validateCarEngineVolume(double value) {
-        if (value <= 0){return 1.5;}
-        return value;
-    }
-    public static int validateCarProductionYear(int value) {
-        if (value <= 0){return 2000;}
-        return value;
-    }
-    public static String validateCarRegistrationNumber(String value) {
-        if (value.matches("[а,в,е,к,м,н,о,р,с,т,у,х]{1}[0-9]{3}[а,в,е,к,м,н,о,р,с,т,у,х]{2}[0-9]{3}")){
-            return value;
-        }
-        return "Некорректный номер";
-    }
-    public static int validateCarNumberOfSeats(int value) {
-        if (value <= 0){return 5;}
-        return value;
-    }
-    public static boolean validateBoolean(boolean value){
-        if (value == true || value == false) {return value;}
-        return false;
-    }
+
     public void changTires(int month){
         if (month >= 4 && month <= 10){
             this.isSummerTires = true;
@@ -171,8 +102,8 @@ public class Car {
         private final boolean keylessAccess;
 
         public Key(boolean remoteEngineStart, boolean keylessAccess) {
-            this.remoteEngineStart = validateBoolean(remoteEngineStart);
-            this.keylessAccess = validateBoolean(keylessAccess);
+            this.remoteEngineStart = remoteEngineStart;
+            this.keylessAccess = keylessAccess;
         }
 
         @Override
